Validati gli input dei REST

+++ src/main/java/com/cybersecurity/progetto_cybersecurity/controller/PostoController.java
@@ -1,6 +1,7 @@
 package com.cybersecurity.progetto_cybersecurity.controller;
 
 import com.cybersecurity.progetto_cybersecurity.controller.dto.PostoDTO;
+import com.cybersecurity.progetto_cybersecurity.security.InputValidator;
 import com.cybersecurity.progetto_cybersecurity.services.PostoService;
 import com.cybersecurity.progetto_cybersecurity.utility.VoloPostoResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,9 @@ public class PostoController {
     @Autowired
     private PostoService postoService;
 
+    @Autowired
+    private InputValidator validator;
+
     @GetMapping
     public List<PostoDTO> getAllPosti() {
         return postoService.getAllPosti();
@@ -25,14 +29,19 @@ public class PostoController {
 
     @GetMapping("/{id}")
     public ResponseEntity<PostoDTO> getPostoById(@PathVariable Long id) {
+        if(!validator.isValidId(id)){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         PostoDTO postoDTO = postoService.getPostoById(id);
         return ResponseEntity.ok(postoDTO);
     }
 
     @GetMapping("/findByCodiceVolo/{codiceVolo}")
-    public ResponseEntity<List<VoloPostoResponse>> getPostoByVolo(@PathVariable String codiceVolo) {
+    public ResponseEntity<List<VoloPostoResponse>> getPostoByVolo(@PathVariable String codiceVolo)
+    {   if(validator.isValidString(codiceVolo)){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         return ResponseEntity.ok(postoService.getPostiByCodiceVolo(codiceVolo));
-
     }
 
     @PostMapping
@@ -44,6 +53,9 @@ public class PostoController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePosto(@PathVariable Long id) {
+        if(!validator.isValidId(id)){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         postoService.deletePosto(id);
         return ResponseEntity.noContent().build();  // Restituisce HTTP 204 (No Content)
     }
